Print usage info when usage is incorrect
Previously we were always silencing usage and not printing
it on error.
This commit silences usage only after the usage is checked
so that we will print usage info on incorrect usage but not
on other errors.

The way this is achieved is by not silencing usage on
the root command and instead setting `SilenceUsage` to `true`
at the beginning of `RunE` method of each leaf command.
This method is executed after the usage is checked for
correctness by Cobra.

Fixes #128
Fixes #132

Signed-off-by: Piotr Jastrzebski <piotr@chiselstrike.com>