small shell QoL improvements
Add a command to show all tables, and another to show all tables
schemas.

This is very basic, and won't even hide the column names in the result
set for those cases. It also doesn't allow to do .schema <something>,
and always shows all tables.

But because we're investing in a proper shell, all I am shooting for
here is some basic ability to do it, otherwise this is too inconvenient.