allow .tables and .schema without interactive shell
This allows passing in `.tables` and `.schema` without calling in to the interactive shell:

```
turso db shell test-db ".tables"
turso db shell test-db ".schema"
```

The checks for those statements were also moved into `query()` so they
would work. Also updates some of the docs to reflect that passing in SQL
as an argument to the `shell` command is possible.