change help text for db replicate
What we are doing now, is printing a replica-specific URL.
This gives users the impression that they may have to connect there
specifically to reach this replica.

Instead, change the text to reinforce that the main URL should work
automatically. Tell them how to acquire a specific URL if they really
need to.