experimental: local development server
Developers using environments like cloudflare or other edge function
providers cannot use the filesystem easily for local testing.
And the ones with a file system usually don't necessarily know
about the file: path.

We have been toying with the idea of spawning a local sqld server, but
it has some disadvantages, mainly it is a big (30MB) Rust binary that
needs to be downloaded separately.

We don't need any of the replication bits from sqld, so it may be
easier to just spawn a webserver and use a local libsql or sqlite
file and proxy the requests.

We'd have to have a hrana-server implementation in go, but that doesn't
sound too hard.

A bit harder, is making sure that all the responses and error messages
are the same as sqld would generate, and keeping that in sync.

This PR adds a very basic wrapper that shows how the devex could be.
I didn't worry about returning any valid responses yet, just execute a
statement, print it, and done. A first step!