Replace sqlite3dump with directly using sqlite3 from $PATH
sqlite3dump requires turso-cli to be built with CGO. Even though it
works when you you build the CLI locally (if you have CGO enabled),
it won't work for our official release builds. Compiling multiplatform
binaries with CGO isn't trivial and it is not worth it. sqlite3 should
be available on most systems and it is reasonable to depend on it for
features such as --from-file.