auth login: improve browser failure error handling
If we run `turso auth login`, let's fail more gracefully if we fail to
open the browser.

Without this change, you'll see:

```
$ turso auth login
error: Unable to open browser.
Visit this URL on this device to log in:
https://api.turso.io?port=39935&redirect=true
Waiting for authentication...
```

With the change, it's now:

```
$ turso auth login
Error: failed to open browser, please use the `--headless` command line option.
```