Use persistent connection for shell
This patch uses websocket connections instead of
HTTP for the shell. In HTTP mode, we use a connection
pool. So, at every invocation in shell, can use a
different connection. While this is okay for most cases,
but this doesn't work for operations involving stateful
connections viz. transactions, ATTACH statements and a
few PRAGMA statements.

closes:
https://github.com/tursodatabase/turso-cli/issues/784
https://github.com/tursodatabase/libsql-shell-go/issues/163