add group show command
After group update, it is more natural to recommend group show to check
the version, versus finding a db inside the group, and then checking its
version.

There is no payload in the group API response with the version, so for
now we just iterate over the databases in the group.