fix logic for selecting default group
The current logic calls groupFromFlag to acquire the group name,
but then has a check later to call the group "default" if isDefault
is true.

This logic is incorrect: groupFromFlag itself has a case where we
return an existing group, and that is the default.

As a result, the following sequence of commands fail:

$ turso group create aws  # ok
$ turso db create --group aws aws # ok
$ turso db create --from-db aws aws-fork # fails
Error: could not create database aws-fork: group not found